// types/register-user.ts
export interface RegisterUser {
  name: string;
  email: string;
  title: string;
  day: string;
  month: string;
  year: string;
  password: string;
  firstName: string;
  lastName: string;
  company?: string;
  address1: string;
  address2?: string;
  country: string;
  city: string;
  state: string;
  zipCode: string;
  mobileNumber: string;
}